Obtain interface MAC from ifreq
An interfaces MAC address can be obtained from the getifaddrs ifreq
list. This saves us performing an extra ioctl to query the interface's
own MAC address and allows us to get the correct MAC address length on
the resepective interface under the assumption that all MAC addresses
on a given interface have the same type and length, as struct arpreq
contains neither a type nor a length.